+++ domain/gateway/src/main/java/app/vahid/domain/gateway/model/Transaction.kt
@@ -1,10 +1,11 @@
 package app.vahid.domain.gateway.model
 
+import java.math.BigDecimal
 
 data class Transaction(
     val originCurrency: String,
     val destinationCurrency: String,
-    val originAmount: Double,
-    val destinationAmount: Double,
+    val originAmount: BigDecimal,
+    val destinationAmount: BigDecimal,
     val fee: Double,
 )
