+++ repository/src/main/java/app/vahid/repository/datasource/CurrencyExchangeRemoteDataSource.kt
@@ -2,11 +2,12 @@ package app.vahid.repository.datasource
 
 import app.vahid.common.core.WrappedResult
 import app.vahid.repository.model.CurrencyRateListEntity
+import java.math.BigDecimal
 
 interface CurrencyExchangeRemoteDataSource {
 
     fun exchangeCurrency(
-        amount: Double,
+        amount: BigDecimal,
         currencyId: String,
         baseCurrencyId: String,
     ): WrappedResult<Unit>
