+++ repository/src/main/java/app/vahid/repository/model/TransactionEntity.kt
@@ -1,9 +1,11 @@
 package app.vahid.repository.model
 
+import java.math.BigDecimal
+
 data class TransactionEntity(
     val originCurrency: String,
     val destinationCurrency: String,
-    val originAmount: Double,
-    val destinationAmount: Double,
+    val originAmount: BigDecimal,
+    val destinationAmount: BigDecimal,
     val fee: Double,
 )
